README.md: Add URL file header
There's already a X-URL (whose purpose is unknown to the author of this
patch), but it's not understood by make-readme-markdown, so add the
somewhat redundant URL file header.

With this we can also remove the melpa badges from the Commentary section
since they will get picked up automatically by make-readme-markdown.